# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/admin/transactions/{id}/status": {
            "patch": {
                "description": "Update status of transaction (admin)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Update transaction status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Status Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateTransactionStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
...
        "/transactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all transactions for authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get user transactions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetTransactionResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create transaction with products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Create new transaction",
                "parameters": [
                    {
                        "description": "Create Transaction Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/transactions/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get transaction detail by ID (user only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get transaction detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetTransactionDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Transaction not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.CreateTransactionRequest": {
            "type": "object",
            "required": [
                "address",
                "delivery_fee",
                "delivery_method",
                "email",
                "full_name",
                "items",
                "subtotal_price",
                "tax",
                "total_price"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "delivery_fee": {
                    "type": "integer"
                },
                "delivery_method": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TransactionItemRequest"
                    }
                },
                "payment_method": {
                    "type": "string"
                },
                "subtotal_price": {
                    "type": "integer"
                },
                "tax": {
                    "type": "integer"
                },
                "total_price": {
                    "type": "integer"
                }
            }
        },
        "dto.CreateTransactionResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/dto.CreateTransactionResult"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "dto.CreateTransactionResult": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "transaction_code": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetTransactionDetailResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/dto.TransactionDetailResponse"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "dto.GetTransactionResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TransactionResult"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Users"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.TransactionDetailResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "delivery_fee": {
                    "type": "integer"
                },
                "delivery_method": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TransactionItemResponse"
                    }
                },
                "payment_method": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "subtotal_price": {
                    "type": "integer"
                },
                "tax": {
                    "type": "integer"
                },
                "total_price": {
                    "type": "integer"
                },
                "transaction_code": {
                    "type": "string"
                }
            }
        },
        "dto.TransactionItemRequest": {
            "type": "object",
            "required": [
                "price",
                "product_id",
                "qty",
                "size",
                "variant"
            ],
            "properties": {
                "price": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "qty": {
                    "type": "integer",
                    "minimum": 1
                },
                "size": {
                    "type": "string"
                },
                "variant": {
                    "type": "string"
                }
            }
        },
        "dto.TransactionItemResponse": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "qty": {
                    "type": "integer"
                },
                "size": {
                    "type": "string"
                },
                "variant": {
                    "type": "string"
                }
            }
        },
        "dto.TransactionResult": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "delivery_fee": {
                    "type": "string"
                },
                "delivery_method": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "payment_method": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "subtotal_price": {
                    "type": "integer"
                },
                "tax": {
                    "type": "integer"
                },
                "total_price": {
                    "type": "integer"
                },
                "transaction_code": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateTransactionStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.Users": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}